Solun 2

+++ src/main/java/com/driver/MovieRepository.java
@@ -2,10 +2,7 @@ package com.driver;
 
 import org.springframework.stereotype.Repository;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 @Repository
 public class MovieRepository {
@@ -14,7 +11,7 @@ public class MovieRepository {
 
     HashMap<String, Director> directorDb = new HashMap<>();
 
-    HashMap<String, String> movieDirectorDb = new HashMap<>();
+    HashMap<String, List<String>> movieDirectorDb = new HashMap<>();
 
     public void addMovie(Movie movie) {
         String key = movie.getName();
@@ -27,7 +24,18 @@ public class MovieRepository {
     }
 
     public void addMovieDirectorPair(String movie, String director) {
-        movieDirectorDb.put(movie, director);
+
+        if(movieDb.containsKey(movie) && directorDb.containsKey(director)){
+            List<String> currentMoviesByDirector = new ArrayList<>();
+
+            if(movieDirectorDb.containsKey(director)) {
+                currentMoviesByDirector = movieDirectorDb.get(director);
+            }
+                currentMoviesByDirector.add(movie);
+
+                movieDirectorDb.put(movie, currentMoviesByDirector);
+        }
+
     }
 
     public Movie getMovieByName(String name) {
@@ -40,42 +48,67 @@ public class MovieRepository {
 
     public List<String> getMoviesByDirectorName(String directorName) {
         List<String> movies = new ArrayList<>();
-        for (Map.Entry<String, String> entry : movieDirectorDb.entrySet()) {
-            if(entry.getValue().equals(directorName)){
-                String movie = entry.getKey();
-                movies.add(movie);
+            if(movieDirectorDb.containsKey(directorName)){
+                movies = movieDirectorDb.get(directorName);
             }
-        }
         return movies;
     }
 
     public List<String> findAllMovies() {
-        List<String> listOfMovies = new ArrayList<>();
-        for (String movieNames : movieDb.keySet()){
-            listOfMovies.add(movieNames);
-        }
-        return listOfMovies;
+        return new ArrayList<>(movieDb.keySet());
+//        List<String> listOfMovies = new ArrayList<>();
+//        for (String movieNames : movieDb.keySet()){
+//            listOfMovies.add(movieNames);
+//        }
+//        return listOfMovies;
     }
 
     public void deleteDirectorByName(String director) {
-        directorDb.remove(director);
-        for(Map.Entry<String,String> entry : movieDirectorDb.entrySet()){
-            if(entry.getValue().equals(director)){
-                String movie = entry.getKey();
-                movieDb.remove(movie);
-                movieDirectorDb.remove(movie,director);
+        List<String> movies = new ArrayList<String>();
+
+        if (movieDirectorDb.containsKey(director)) {
+            //1. Find the movie names by director from the pair
+            movies = movieDirectorDb.get(director);
+
+            //2. Deleting all the movies from movieDb by using movieName
+            for (String movie : movies) {
+                if (movieDb.containsKey(movie)) {
+                    movieDb.remove(movie);
+                }
             }
+
+            //3. Deleteing the pair
+            movieDirectorDb.remove(director);
         }
 
+        //4. Delete the director from directorDb.
+        if(directorDb.containsKey(director)){
+            directorDb.remove(director);
+        }
     }
 
     public void deleteAllDirectors() {
-        for (Map.Entry<String,String> entry : movieDirectorDb.entrySet()) {
-            String movie = entry.getKey();
-            String director = entry.getValue();
-            movieDirectorDb.remove(movie,director);
-            movieDb.remove(movie);
-            directorDb.remove(director);
+        HashSet<String> moviesSet = new HashSet<String>();
+
+        //Deleting the director's map
+        directorDb = new HashMap<>();
+
+        //Finding out all the movies by all the directors combined
+        for(String director: movieDirectorDb.keySet()){
+
+            //Iterating in the list of movies by a director.
+            for(String movie: movieDirectorDb.get(director)){
+                moviesSet.add(movie);
+            }
         }
+
+        //Deleting the movie from the movieDb.
+        for(String movie: moviesSet){
+            if(movieDb.containsKey(movie)){
+                movieDb.remove(movie);
+            }
+        }
+        //clearing the pair.
+        movieDirectorDb = new HashMap<>();
     }
 }
